Skip rabbitmq secret creation if already exist
The operator will continue to run reconciliation with the playbook
as with each run the status with changed is not 0. In the Operand
status condition the status will be always running without show
ansibleResult.

In ReportPortal depoly this is caused by the rabbitmq-erlang-cookie
value changed with each run, which could be skipped if the secret
have already been created.

Signed-off-by: Wayne Sun <gsun@redhat.com>